requirements.txt
torch>=2.0
torchvision
numpy
Pillow
pytorch-fid
wandb
datasets